+++ tests/test_bootstrap_field_input.py
@@ -6,21 +6,37 @@ from django.test import TestCase
 from .base import render_template_with_bootstrap
 
 
-class TestForm(forms.Form):
+class TextTestForm(forms.Form):
     test = forms.CharField()
 
 
-class BootstrapFieldTest(TestCase):
+class FileFieldTestForm(forms.Form):
+    test = forms.FileField()
+
+
+class ClearableFileInputTestForm(forms.Form):
+    test = forms.FileField(widget=forms.ClearableFileInput, required=False)
+
+
+class CheckboxTestForm(forms.Form):
+    test = forms.BooleanField()
+
+
+class ColorTestForm(forms.Form):
+    test = forms.CharField(widget=TextInput(attrs={"type": "color"}))
+
+
+class RangeTestForm(forms.Form):
+    test = forms.IntegerField(widget=TextInput(attrs={"type": "range"}))
+
+
+class InputTypeTextTestCase(TestCase):
     def test_input_type_text(self):
         """Test field with text widget."""
 
-        class TestForm(forms.Form):
-            test = forms.CharField()
-
-        test_form = TestForm()
-        html = render_template_with_bootstrap("{% bootstrap_field form.test %}", context={"form": test_form})
+        test_form = TextTestForm()
         self.assertHTMLEqual(
-            html,
+            render_template_with_bootstrap("{% bootstrap_field form.test %}", context={"form": test_form}),
             (
                 '<div class="django_bootstrap5-req mb-3">'
                 '<label for="id_test" class="form-label">Test</label>'
@@ -32,12 +48,11 @@ class BootstrapFieldTest(TestCase):
     def test_bootstrap_field_text_horizontal(self):
         """Test field with text widget."""
 
-        test_form = TestForm()
-        html = render_template_with_bootstrap(
-            "{% bootstrap_field form.test layout='horizontal' %}", context={"form": test_form}
-        )
+        test_form = TextTestForm()
         self.assertHTMLEqual(
-            html,
+            render_template_with_bootstrap(
+                "{% bootstrap_field form.test layout='horizontal' %}", context={"form": test_form}
+            ),
             (
                 '<div class="django_bootstrap5-req mb-3 row">'
                 '<label class="col-form-label col-sm-2" for="id_test">'
@@ -52,15 +67,11 @@ class BootstrapFieldTest(TestCase):
     def test_input_type_text_floating(self):
         """Test field with text widget in floating layout."""
 
-        class TestForm(forms.Form):
-            test = forms.CharField()
-
-        test_form = TestForm()
-        html = render_template_with_bootstrap(
-            "{% bootstrap_field form.test layout='floating' %}", context={"form": test_form}
-        )
+        test_form = TextTestForm()
         self.assertHTMLEqual(
-            html,
+            render_template_with_bootstrap(
+                "{% bootstrap_field form.test layout='floating' %}", context={"form": test_form}
+            ),
             (
                 '<div class="django_bootstrap5-req mb-3 form-floating">'
                 '<input class="form-control" id="id_test" name="test" placeholder="Test" required type="text">'
@@ -69,16 +80,14 @@ class BootstrapFieldTest(TestCase):
             ),
         )
 
+
+class InputTypeColorTestCase(TestCase):
     def test_input_type_color(self):
-        """Test field with text widget."""
+        """Test field with input widget with type `color`."""
 
-        class TestForm(forms.Form):
-            test = forms.CharField(widget=TextInput(attrs={"type": "color"}))
-
-        test_form = TestForm()
-        html = render_template_with_bootstrap("{% bootstrap_field form.test %}", context={"form": test_form})
+        test_form = ColorTestForm()
         self.assertHTMLEqual(
-            html,
+            render_template_with_bootstrap("{% bootstrap_field form.test %}", context={"form": test_form}),
             (
                 '<div class="django_bootstrap5-req mb-3">'
                 '<label for="id_test" class="form-label">Test</label>'
@@ -88,16 +97,101 @@ class BootstrapFieldTest(TestCase):
             ),
         )
 
-    def test_input_type_checkbox(self):
-        """Test field with text widget."""
+    def test_input_type_color_horizontal(self):
+        """Test field with input widget with type `color` in horizontal layout."""
 
-        class TestForm(forms.Form):
-            test = forms.BooleanField()
-
-        test_form = TestForm()
-        html = render_template_with_bootstrap("{% bootstrap_field form.test %}", context={"form": test_form})
+        test_form = ColorTestForm()
         self.assertHTMLEqual(
-            html,
+            render_template_with_bootstrap(
+                '{% bootstrap_field form.test layout="horizontal" %}', context={"form": test_form}
+            ),
+            (
+                '<div class="django_bootstrap5-req row mb-3">'
+                '<label for="id_test" class="col-form-label col-sm-2">Test</label>'
+                '<div class="col-sm-10">'
+                '<input class="form-control form-control-color" id="id_test"'
+                ' name="test" placeholder="Test" required type="color">'
+                "</div>"
+                "</div>"
+            ),
+        )
+
+    def test_input_type_color_floating(self):
+        """Test field with input widget with type `color` in floating layout."""
+
+        test_form = ColorTestForm()
+        self.assertHTMLEqual(
+            render_template_with_bootstrap(
+                '{% bootstrap_field form.test layout="floating" %}', context={"form": test_form}
+            ),
+            (
+                '<div class="django_bootstrap5-req mb-3">'
+                '<label for="id_test" class="form-label">Test</label>'
+                '<input class="form-control form-control-color" id="id_test"'
+                ' name="test" placeholder="Test" required type="color">'
+                "</div>"
+            ),
+        )
+
+
+class InputTypeRangeTestCase(TestCase):
+    def test_input_type_range(self):
+        """Test field with input widget with type `range`."""
+
+        test_form = RangeTestForm()
+
+        self.assertHTMLEqual(
+            render_template_with_bootstrap("{% bootstrap_field form.test %}", context={"form": test_form}),
+            (
+                '<div class="django_bootstrap5-req mb-3">'
+                '<label for="id_test" class="form-label">Test</label>'
+                '<input class="form-range" id="id_test" name="test" placeholder="Test" required type="range">'
+                "</div>"
+            ),
+        )
+
+    def test_input_type_range_horizontal(self):
+        """Test field with input widget with type `range` in horizontal layout."""
+
+        test_form = RangeTestForm()
+        self.assertHTMLEqual(
+            render_template_with_bootstrap(
+                '{% bootstrap_field form.test layout="horizontal" %}', context={"form": test_form}
+            ),
+            (
+                '<div class="django_bootstrap5-req row mb-3">'
+                '<label for="id_test" class="col-form-label col-sm-2">Test</label>'
+                '<div class="col-sm-10">'
+                '<input class="form-range" id="id_test" name="test" placeholder="Test" required type="range">'
+                "</div>"
+                "</div>"
+            ),
+        )
+
+    def test_input_type_range_floating(self):
+        """Test field with input widget with type `range` in floating layout."""
+
+        test_form = RangeTestForm()
+        self.assertHTMLEqual(
+            render_template_with_bootstrap(
+                '{% bootstrap_field form.test layout="floating" %}', context={"form": test_form}
+            ),
+            (
+                '<div class="django_bootstrap5-req mb-3">'
+                '<label for="id_test" class="form-label">Test</label>'
+                '<input class="form-range" id="id_test" name="test" placeholder="Test" required type="range">'
+                "</div>"
+            ),
+        )
+
+
+class InputTypeCheckboxTestCase(TestCase):
+    def test_input_type_checkbox(self):
+        """Test field with checkbox widget."""
+
+        test_form = CheckboxTestForm()
+        self.assertHTMLEqual(
+            render_template_with_bootstrap("{% bootstrap_field form.test %}", context={"form": test_form}),
             (
                 '<div class="django_bootstrap5-req mb-3">'
                 '<div class="form-check">'
@@ -109,17 +203,13 @@ class BootstrapFieldTest(TestCase):
         )
 
     def test_input_type_checkbox_style_switch(self):
-        """Test field with text widget."""
+        """Test field with checkbox widget, style switch."""
 
-        class TestForm(forms.Form):
-            test = forms.BooleanField()
-
-        test_form = TestForm()
-        html = render_template_with_bootstrap(
-            '{% bootstrap_field form.test checkbox_style="switch" %}', context={"form": test_form}
-        )
+        test_form = CheckboxTestForm()
         self.assertHTMLEqual(
-            html,
+            render_template_with_bootstrap(
+                '{% bootstrap_field form.test checkbox_style="switch" %}', context={"form": test_form}
+            ),
             (
                 '<div class="django_bootstrap5-req mb-3">'
                 '<div class="form-check form-switch">'
@@ -131,17 +221,13 @@ class BootstrapFieldTest(TestCase):
         )
 
     def test_bootstrap_field_checkbox_horizontal(self):
-        """Test field with text widget."""
+        """Test field with checkbox widget, layout horizontal."""
 
-        class TestForm(forms.Form):
-            test = forms.BooleanField()
-
-        test_form = TestForm()
-        html = render_template_with_bootstrap(
-            "{% bootstrap_field form.test layout='horizontal' %}", context={"form": test_form}
-        )
+        test_form = CheckboxTestForm()
         self.assertHTMLEqual(
-            html,
+            render_template_with_bootstrap(
+                "{% bootstrap_field form.test layout='horizontal' %}", context={"form": test_form}
+            ),
             (
                 '<div class="django_bootstrap5-req mb-3 row">'
                 '<div class="col-sm-10 offset-sm-2">'
@@ -154,14 +240,13 @@ class BootstrapFieldTest(TestCase):
             ),
         )
 
-    def test_input_type_file(self):
-        class TestForm(forms.Form):
-            test = forms.FileField()
 
-        test_form = TestForm()
-        html = render_template_with_bootstrap("{% bootstrap_field form.test %}", context={"form": test_form})
+class InputTypeFileTestCase(TestCase):
+    def test_input_type_file(self):
+
+        test_form = FileFieldTestForm()
         self.assertHTMLEqual(
-            html,
+            render_template_with_bootstrap("{% bootstrap_field form.test %}", context={"form": test_form}),
             (
                 '<div class="django_bootstrap5-req mb-3">'
                 '<label for="id_test" class="form-label">Test</label>'
@@ -171,10 +256,8 @@ class BootstrapFieldTest(TestCase):
         )
 
     def test_clearable_file_input(self):
-        class TestForm(forms.Form):
-            test = forms.FileField(widget=forms.ClearableFileInput, required=False)
 
-        test_form = TestForm()
+        test_form = ClearableFileInputTestForm()
         self.assertHTMLEqual(
             render_template_with_bootstrap("{% bootstrap_field form.test %}", context={"form": test_form}),
             (
@@ -186,10 +269,8 @@ class BootstrapFieldTest(TestCase):
         )
 
     def test_clearable_file_input_post(self):
-        class TestForm(forms.Form):
-            test = forms.FileField(widget=forms.ClearableFileInput, required=False)
 
-        test_form = TestForm({}, {"test": SimpleUploadedFile("test.txt", b"test")})
+        test_form = ClearableFileInputTestForm({}, {"test": SimpleUploadedFile("test.txt", b"test")})
         self.assertHTMLEqual(
             render_template_with_bootstrap("{% bootstrap_field form.test %}", context={"form": test_form}),
             (